Split the enums out of camada.h into camadaenums.h
camada.h is what a consumer reads to learn the API, and 215 of its 1433
lines were enum definitions and their documentation -- the FXPRM comment
alone is 35 lines. Moving them leaves camada.h as the operation surface.

camada.h includes the new header, so it stays a single umbrella include
and no consumer changes. camadaenums.h has no includes of its own: it is
pure enum definitions, which is worth keeping true.

Extracting the block initially swept up getCamadaVersion(), which sits
between two enums in the original -- caught by compiling the new header
on its own, which failed on the std::string it needs. Moved back.
