+++ src/camada.h
@@ -29,226 +29,14 @@
 #include <vector>
 
 #include "camadacommon.h"
+#include "camadaenums.h"
 #include "camadaexpr.h"
 #include "camadasort.h"
 
 namespace camada {
 
-/// Selects how Camada represents floating-point values when constructing
-/// FP/RM sorts and FP constants. The encoding is per-sort, not solver-
-/// wide — a single solver instance can hold both Native and BV-encoded
-/// FP values at the same time, and they interoperate through the
-/// common-layer encoders.
-///
-/// - Native: use the backend's native floating-point sort (`Float32Sort`
-///   in z3, `mkFloatingPoint` in cvc5, etc.). Requires native FP
-///   support in the backend; fastest path on solvers that have it.
-/// - BV: bit-blast every FP value into a fixed-width bit-vector and
-///   emulate the IEEE-754 operations through Camada's common-layer
-///   encoder. The only path available on backends without native FP
-///   (STP, Yices-SMT2) and on SMT-LIB scripts intended for solvers that
-///   reject native FP. Substantially slower than Native on backends
-///   that have both.
-///
-/// The two encodings round-trip cleanly across all five FP arithmetic
-/// ops, predicates, and conversions (the `fp_native_bv_predicate_parity`
-/// regression pins this), but model values are reported in the encoded
-/// representation — `getFP32` decodes BV back to `float`, `getBV`
-/// returns the raw bits when the sort was Native.
-enum class FPEncoding { Native, BV };
-
-/// Selects how `mkArrayConst` lowers a constant array. Unlike `FPEncoding`
-/// this is not a sort property: lazily and natively lowered constant arrays
-/// share the same array sorts and interoperate freely (stores, selects, and
-/// ites may mix them); the choice only affects how the "every element
-/// equals InitValue" semantics are produced.
-///
-/// - Auto: native `((as const ...) v)` when the backend supports it, the
-///   Camada lazy lowering otherwise. The right default for almost all uses.
-/// - Native: force the backend operator; fails on backends without one.
-/// - Lazy: force the Camada lowering — a fresh array symbol whose default
-///   value is asserted on demand at each index the formula observes. Works
-///   at any index width and is the entry point for initializers that have
-///   no backend representation.
-///
-/// Lazily lowered arrays must stay observable by the lowering: storing one
-/// inside another array, placing one in a native tuple, or passing one to
-/// an uninterpreted function is rejected, and capturing one under a
-/// quantifier binder is unsupported (defaults are instantiated as ground
-/// constraints only, so a quantified body can observe uninstantiated
-/// indexes).
-enum class ConstArrayLowering { Auto, Native, Lazy };
-
-/// Selects how Camada lowers `mkFPNeg` for backends whose native FP
-/// implementation diverges from the IEEE-754 sign-bit-flip semantics
-/// some users expect.
-///
-/// - FlipSignBit: always flip the IEEE-754 sign bit, including on NaN
-///   inputs. Matches the behavior of CPU FP units and most language
-///   runtimes. Backed by an explicit bit-blast on solvers whose native
-///   `fp.neg` preserves the NaN payload — see PR #59 for the per-
-///   backend status.
-/// - PreserveNaNPayload: follow the SMT-LIB `fp.neg` definition, which
-///   leaves NaN payloads (including the sign bit) unchanged. Cheaper to
-///   emit on backends that natively implement this semantics.
-enum class FPNegBehavior {
-  FlipSignBit,
-  PreserveNaNPayload,
-};
-
-/// Rounding direction for the fixed-point operations that discard bits.
-///
-/// Separate from `RM` rather than reusing it, because fixed point needs a
-/// mode IEEE-754 has no name for. TR 18037's `roundfx` rounds to nearest
-/// but leaves the halfway direction to the implementation, and LLVM libc
-/// breaks ties toward positive infinity — nearest for every other value.
-/// `RM::ROUND_TO_PLUS_INF` is a different operation: it rounds *all*
-/// values up, so 0.25 becomes 1 where nearest-ties-up leaves it 0. With no
-/// RM member denoting "nearest, ties toward +inf", reusing RM would have
-/// silently dropped the one `roundfx` direction verified against an
-/// executing implementation.
-///
-/// The three nearest modes differ only on exact halfway values:
-///
-/// - NearestTiesTowardPositive: 0.5 rounds to 1 and -0.5 rounds to 0.
-///   LLVM libc's `roundfx` (it adds half an ulp and masks off the low
-///   bits), and the only direction verified against a running libc.
-/// - NearestTiesAwayFromZero: 0.5 rounds to 1 and -0.5 rounds to -1 —
-///   symmetric about zero, what most C programmers expect from `round()`.
-///   Equivalent to `RM::ROUND_TO_AWAY`.
-/// - NearestTiesToEven: a halfway value rounds to whichever neighbour has
-///   a zero in the last kept bit, IEEE-754's default and the unbiased
-///   choice. Equivalent to `RM::ROUND_TO_EVEN`.
-///
-/// The directed modes never consult the halfway point:
-///
-/// - TowardZero: truncate. C's fixed-to-integer and float-to-fixed
-///   direction, pinned by the execution oracle.
-/// - TowardNegative: floor. What Clang's `-ffixed-point` division does —
-///   verified: -0.5 / 0.75 gives -0.671875, not -0.6640625.
-/// - TowardPositive: ceiling.
-///
-/// Every nearest mode saturates to the format's maximum when the rounding
-/// would carry past it, which every implementation surveyed agrees on.
-enum class FXPRM {
-  NearestTiesTowardPositive,
-  NearestTiesAwayFromZero,
-  NearestTiesToEven,
-  TowardZero,
-  TowardNegative,
-  TowardPositive,
-};
-
-/// Selects how the SMT-LIB backend lowers tuples on the wire.
-///
-/// - Native: emit `(declare-datatypes ...)` and rely on the downstream
-///   solver to support SMT-LIB datatypes. Works against z3 and cvc5; not
-///   accepted by bitwuzla, mathsat, yices-smt2.
-/// - Camada: lower tuples in Camada to per-field BV/Bool symbols before
-///   anything reaches the wire. The emitted script contains no
-///   datatype declarations, so any standard SMT-LIB v2 solver can parse
-///   it. Same encoding the non-native backends (bitwuzla/mathsat/stp/
-///   yices) already use.
-enum class TupleEncoding { Native, Camada };
-
-/// Selects whether a solver context is created with unsat-assumption
-/// production enabled. Producing the core is not free: backends whose SAT
-/// engine must track assumption participation (bitwuzla, cvc5) pay a
-/// solve-time cost on *every* check, and the setting is frozen at context
-/// creation — so it is opt-in.
-///
-/// - Off (default): fast contexts. checkSatAssuming() works unchanged;
-///   getUnsatAssumptions() reports UnsupportedOperation and
-///   supports(SolverFeature::UnsatAssumptions) answers false.
-/// - On: the backend tracks assumptions and getUnsatAssumptions() returns
-///   real cores after an UNSAT checkSatAssuming().
-///
-/// Backends that answer cores without a creation-time option (Z3 enables
-/// unsat_core per query, MathSAT and Yices track natively) ignore this
-/// and always support core extraction.
-enum class UnsatAssumptionsMode { Off, On };
-
-/// Selects how arrays are encoded, on the backends that accept it.
-///
-/// - Native (default): the backend's theory of arrays, unchanged.
-/// - Ackermann: arrays never reach the backend. Every select becomes a
-///   fresh element variable tied to the array's other reads by congruence
-///   axioms (`i = j => a[i] = a[j]`); stores and ites are lowered
-///   structurally, and array equality uses a witness-index encoding. The
-///   trade: array-theory work moves into the core solver as ground
-///   constraints, quadratic in the number of reads per array.
-///
-/// Restrictions in Ackermann mode: quantifier-free formulas only (any
-/// mkForall/mkExists call is rejected), no nested arrays, no array-sorted
-/// UF arguments/returns, and model queries need bool/BV index sorts. The
-/// mode forces the Camada tuple encoding — a native datatype cannot hold
-/// an array member that has no backend representation.
-enum class ArrayEncoding { Native, Ackermann };
-
-enum class RM {
-  ROUND_TO_EVEN = 0,
-  ROUND_TO_AWAY = 1,
-  ROUND_TO_PLUS_INF = 2,
-  ROUND_TO_MINUS_INF = 3,
-  ROUND_TO_ZERO = 4,
-};
-
 /// Return camada version
 std::string getCamadaVersion();
-
-enum class checkResult { SAT, UNSAT, UNKNOWN };
-
-/// Capabilities a backend may or may not implement, queryable through
-/// SMTSolver::supports() instead of discovering them through aborts or
-/// UnsupportedOperation errors.
-///
-/// A true bit means the corresponding API surface is implemented for the
-/// backend; individual calls can still fail for input-specific reasons
-/// through their SMTResult. On the SMT-LIB pipeline backend the bits
-/// describe what Camada emits — a particular child solver may still
-/// reject a construct at runtime.
-enum class SolverFeature {
-  /// Int/Real sorts and arithmetic (mkIntSort, mkRealSort, mkArith*).
-  IntRealArithmetic,
-  /// Quantified formulas (mkForall, mkExists).
-  Quantifiers,
-  /// Uninterpreted functions (mkFunctionSort, mkApply).
-  UninterpretedFunctions,
-  /// FPEncoding::Native sorts and operations; FPEncoding::BV works on
-  /// every backend regardless.
-  NativeFloatingPoint,
-  /// Backend-native tuple/datatype sorts; other backends route tuples
-  /// through the Camada per-field lowering.
-  NativeTuples,
-  /// Backend-native `((as const ...) v)` constant arrays; other backends
-  /// lower them lazily (see ConstArrayLowering).
-  NativeConstantArrays,
-  /// Unsat-assumption extraction after an UNSAT checkSatAssuming()
-  /// (see issue #76). checkSatAssuming itself works on every backend
-  /// through a push/assert/check/pop fallback.
-  UnsatAssumptions,
-  /// Per-check wall-clock limits via setTimeout() (see issue #77).
-  Timeouts,
-  /// Sparse array model extraction via getArrayValues() for arbitrary
-  /// arrays (see issue #79). Lazily lowered constant arrays are answered
-  /// by the common layer on every backend regardless.
-  ArrayModels,
-};
-
-/// Coarse-grained error categories for operations that return `SMTResult<T>`.
-///
-/// These are intended for user-triggerable failures such as unsupported
-/// features or model-query failures, not internal invariant violations.
-enum class SMTErrorCode {
-  None,
-  BackendError,
-  InvalidModelValue,
-  UnsupportedOperation,
-  /// The call sequence violated an API contract (e.g. querying unsat
-  /// assumptions after the solver state changed), as opposed to a
-  /// backend-originated failure.
-  InvalidUsage,
-};
 
 /// Structured error payload carried by `SMTResult<T>` on failure.
 struct SMTError {
